seperated upload and trip image component
