+++ client/app/singletrip/page.js
@@ -21,6 +21,7 @@ import TripIconBarComponent from '../components/singletrip/TripIconBarComponent'
 import BarGraphComponent from '../components/singletrip/BarGraphComponent';
 import SpendingCirclesComponent from '../components/singletrip/SpendingCirclesComponent';
 import TripImageComponent from '../components/singletrip/TripImageComponent';
+import UploadTripImage from '../components/singletrip/UploadTripImage';
 
 
 
@@ -504,14 +505,7 @@ function Singletrip() {
 
 
                                 {/* Add Image Button */}
-                                <div className="icon-div" tooltip="Add Image" tabIndex="0">
-                                    <div className="icon-SVG">
-                                        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth="1.3" stroke="currentColor" className="size-6">
-                                            <path strokeLinecap="round" strokeLinejoin="round" d="m2.25 15.75 5.159-5.159a2.25 2.25 0 0 1 3.182 0l5.159 5.159m-1.5-1.5 1.409-1.409a2.25 2.25 0 0 1 3.182 0l2.909 2.909m-18 3.75h16.5a1.5 1.5 0 0 0 1.5-1.5V6a1.5 1.5 0 0 0-1.5-1.5H3.75A1.5 1.5 0 0 0 2.25 6v12a1.5 1.5 0 0 0 1.5 1.5Zm10.5-11.25h.008v.008h-.008V8.25Zm.375 0a.375.375 0 1 1-.75 0 .375.375 0 0 1 .75 0Z" />
-                                        </svg>
-                                        <span className="icon-text">Add Image</span>
-                                    </div>
-                                </div>
+                                <UploadTripImage tripId={tripId} />
 
                                 {/* Applied filter popup */}
                                 {selectedFilter && (
